Add contribution guidelines
I've added a _Contributing_ section to the README and a [CONTRIBUTING
file](https://github.com/blog/1184-contributing-guidelines). I do not
expected for these changes to be merged right away, as they can be seen
as a draft from what I've observed in the past weeks. The intention here
is not only help the developers that are new to the project, but also
make it easier for Anders to deal with new patches.