Split GCE manifest into regular & backports
Some people want the security team-supported 3.2 kernel, while others
want the better performance of the backports kernel. Both flavors are
currently provided in the Google Compute Engine debian-cloud project.

Also add OpenSSH to the backports flavor, for parity with the current
GCE backports image based on build-debian-cloud. This backport is
included for performance reasons, in line with the purpose of that image
flavor, and as discussed on the debian-cloud list.